change data type (double -> bigdecial; Date -> LocalDateTime)

+++ src/main/java/com/danmin/home_service/dto/request/BookingDTO.java
@@ -27,7 +27,7 @@ public class BookingDTO implements Serializable {
 
     private String workLoad;
 
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 
     @NotNull(message = "type must be not null")
     @EnumPattern(name = "booking_status", regexp = "pending|assigned|in_progress|completed|cancelled|rescheduled")
